requirements.txt
Flask>=2.3.3
gunicorn>=21.2.0
torch>=2.5.0
torchvision>=0.22.0
numpy>=1.26.4
Pillow>=10.3.0
tqdm>=4.66.4
